requirements.txt
streamlit==1.55.0
langchain==1.2.15
langchain-core==1.2.22
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-pinecone==0.2.13
langsmith==0.7.26
langgraph==1.1.6
pinecone==7.3.0
fastapi==0.135.2
uvicorn==0.42.0
motor==3.7.1
pymongo==4.16.0
python-dotenv==1.2.2
requests==2.32.5
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose==3.5.0
unstructured==0.22.23
pdfminer==20191125
faster-whisper==1.2.1
torch==2.11.0
